Revert #5 and #8
Unfortunately, starting from Godot 4.3, we can no longer include empty classes in extensions. Therefore, having the Depth tool extend from BaseDraw is no longer possible, at least in the current state of the API. We either need to change Pixelorama's tools system to accept tools without scenes, that generate their nodes completely via code, or remake the functionality from BaseDraw and add it to the Depth tool.

It's a shame to have to undo these changes, but we can always re-add them in the future.